move urlFilterToNeDbFilter to lib
